pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mercury-ocip"
version = "v0.0.6-beta"
description = "Python SDK for BroadWorks OCI (Open Client Interface) API"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    { name = "Jordan Prescott", email = "devteam@fourteenip.com" },
    { name = "Michael Clarke", email = "devteam@fourteenip.com" },
    { name = "Billy Kelly", email = "devteam@fourteenip.com" },
]
classifiers = [
    # Development Status
    "Development Status :: 3 - Beta",
    # Target Audience
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    # License
    "License :: OSI Approved :: MIT License",
    # Programming Languages
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    # Operating Systems
    "Operating System :: OS Independent",
    # Topics
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Telephony",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
keywords = [
    "broadworks",
    "oci",
    "soap",
    "telephony",
    "voip",
    "sdk",
    "api",
    "mercury",
]
dependencies = [
    "asyncio>=3.4.3",
    "attrs>=21.4.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "requests>=2.32.3",
    "zeep>=4.3.1",
    "types-lxml>=2025.3.30",
    "lxml-stubs>=0.5.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["src"]
addopts = "-ra -q"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]

[tool.ruff]
exclude = ["tests/*", "test_*.py"]
[tool.ruff.lint.per-file-ignores]
"**/oci_requests.py" = ["F821", "F401", "F403", "F405"]
"**/oci_responses.py" = ["F821", "F401", "F403", "F405"]
"**/oci_types.py" = ["F821", "F401", "F403", "F405"]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.15",
    "mkdocstrings[python]>=0.29.1",
    "pymdown-extensions>=10.16",
    "pyrefly>=0.28.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.8.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-material>=9.6.16",
    "mkdocs-minify-html-plugin>=0.3.5",
    "mkdocstrings[python]>=0.30.0",
    "pymdown-extensions>=10.16.1",
]

[tool.pyrefly]
project-excludes = ["tests/", "sandbox.py", "*/commands.py", "docs/"]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
publish-url = "https://upload.pypi.org/legacy/"
